# Backend/static/swagger_config.json
{
    "swagger": "2.0",
    "info": {
        "title": "PlayStation 5 E-commerce API",
        "description": "API documentation for the PlayStation 5 e-commerce website",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": ["http"],
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter 'Bearer' [space] and then your token in the text input below.\n\nExample: 'Bearer abcdef12345'"
        }
    },
    "paths": {
        "/users/register": {
            "post": {
                "tags": ["Users"],
                "summary": "Register a new user",
                "description": "Handles the registration of new users by validating the provided data and creating a new user account.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User object that needs to be registered",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": ["first_name", "last_name", "email", "password"],
                            "properties": {
                                "first_name": { "type": "string" },
                                "last_name": { "type": "string" },
                                "email": { "type": "string" },
                                "password": { "type": "string" }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Registration",
                        "schema": { "type": "string" }
                    },
                    "403": {
                        "description": "Validation error",
                        "schema": {
                            "type": "array",
                            "items": { "type": "string" }
                        }
                    },
                    "409": {
                        "description": "Email already registered",
                        "schema": { "type": "string" }
                    },
                    "400": {
                        "description": "Registration Failed",
                        "schema": { "type": "string" }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "tags": ["Users"],
                "summary": "Login user",
                "description": "Authenticates the user using provided credentials and returns JWT tokens if the login is successful.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User credentials",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": ["email", "password"],
                            "properties": {
                                "email": { "type": "string" },
                                "password": { "type": "string" }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Login successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": { "type": "integer", "format": "int64" },
                                "email": { "type": "string" },
                                "first_name": { "type": "string" },
                                "last_name": { "type": "string" },
                                "token": {
                                    "type": "object",
                                    "properties": {
                                        "access": { "type": "string" },
                                        "refresh": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Validation error or login failed",
                        "schema": { "type": "string" }
                    }
                }
            }
        },
        "/users/update/{pk}": {
            "put": {
                "tags": ["Users"],
                "summary": "Update user information",
                "description": "Allows authenticated users to update their profile information.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "pk",
                        "type": "integer",
                        "required": true,
                        "description": "User ID"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User object that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": ["first_name", "last_name", "email"],
                            "properties": {
                                "first_name": { "type": "string" },
                                "last_name": { "type": "string" },
                                "email": { "type": "string" }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": { "type": "integer", "format": "int64" },
                                "first_name": { "type": "string" },
                                "last_name": { "type": "string" },
                                "email": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Validation error or update failed",
                        "schema": { "type": "string" }
                    }
                },
                "security": [{ "Bearer": [] }]
            }
        },
        "/users/refresh": {
            "post": {
                "tags": ["Users"],
                "summary": "Refresh tokens",
                "description": "Generates new access and refresh tokens for authenticated users by blacklisting the current tokens.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Current tokens",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": ["access", "refresh"],
                            "properties": {
                                "access": { "type": "string" },
                                "refresh": { "type": "string" }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tokens refreshed successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "access": { "type": "string" },
                                "refresh": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Validation error or token refresh failed",
                        "schema": { "type": "string" }
                    }
                },
                "security": [{ "Bearer": [] }]
            }
        },
        "/users/logout": {
            "post": {
                "tags": ["Users"],
                "summary": "Logout user",
                "description": "Logs out the authenticated user by blacklisting the current access and refresh tokens.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Current tokens",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": ["access", "refresh"],
                            "properties": {
                                "access": { "type": "string" },
                                "refresh": { "type": "string" }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Validation error or logout failed",
                        "schema": { "type": "string" }
                    }
                },
                "security": [{ "Bearer": [] }]
            }
        },
...
    }
}